Add test dispatchers
This adds a test implementation of dispatchers with an injected TestDispatcher.
The normal IO dispatcher is automatically swapped out via a TestInstallsIn in core-testing.

This update also updates DataStore to use an injected dispatcher, which allows removing the waitUntil in the navigation tests because we can eagerly read from disk with runTest.

Change-Id: I82abb8c8a57c2beed5ef37bf1e252acf379278cd