Add loading indicator while for you data is loading
Updates the for you view model to support displaying a loading
state when data hasn't loaded yet.

This is a somewhat involved CL, because there are multiple ways
in which data could be loading: We could have topics, but
the feed itself could be loading after we selected a new set
of topics.

Change-Id: I8662140c7132b195f85e69fee8e18745829ae975