Add `ValidationUtils` and update ProGuard rules
This commit introduces a new `ValidationUtils.kt` file with email validation logic.
It also updates the ProGuard rules in `consumer-rules.pro` to keep runtime visible annotations and fields for `EventBus`.
Additionally, a `.gitignore` file is added to the `mylibrary` module to ignore the `build` directory, and the release build type in `app/build.gradle.kts` is now debuggable.